More updates
* Updated README
* renamed Rules to RulesParser
* renamed Rules.fetch_parser to RulesParser.fetch
* implemented behaviour callback for RulesParser
