+++ lib/live_view_native/stylesheet/sheet_parser.ex
@@ -21,4 +21,19 @@ defmodule LiveViewNative.Stylesheet.SheetParser do
       {["color-rainbow", {:_target, [], Elixir}], "red\norange\nyellow\nblue\nindigo\nviolet"}
     ]
   end
+
+  defmacro sigil_SHEET(sheet, _modifier) do
+    blocks =
+      sheet
+      |> LiveViewNative.Stylesheet.Utils.eval_quoted()
+      |> LiveViewNative.Stylesheet.SheetParser.parse()
+
+    quote bind_quoted: [blocks: Macro.escape(blocks)] do
+      for {arguments, body} <- blocks do
+        def class(unquote_splicing(arguments)) do
+          sigil_RULES(unquote(body), nil)
+        end
+      end
+    end
+  end
 end
